equipment_project/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autoflake==2.2.0
black==23.7.0
click==8.1.7
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
protobuf==4.21.12
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
sqlparse==0.4.4
uritemplate==4.1.1
